Wrap for the latest symbol version
GOTCHA previously did not honor symbol versions used in glibc. For example, when wrapping the `pthread_cond_*` family of functions, there are two symbol versions in glibc:

    - `GLIBC_2.2.5`
    - `GLIBC_2.3.2`

Unfortunately, for glibc 2.31 (Ubuntu 20.04), GOTCHA wraps `pthread_cond_init` to `__pthread_cond_init_2_0` but `pthread_cond_broadcast` to `__pthread_cond_broadcast`. I.e., one to the old ABI and the other to the new ABI. The reason appears to be the ordering of the symbols in `libpthread-2.31.so`:

```
   267: 0000000000010610    31 FUNC    GLOBAL DEFAULT   16 pthread_cond_init@GLIBC_2.2.5
   268: 000000000000ed90    58 FUNC    GLOBAL DEFAULT   16 pthread_cond_init@@GLIBC_2.3.2
   205: 0000000000010290   894 FUNC    GLOBAL DEFAULT   16 pthread_cond_broadcast@@GLIBC_2.3.2
   206: 00000000000107c0   101 FUNC    GLOBAL DEFAULT   16 pthread_cond_broadcast@GLIBC_2.2.5
```